pyproject.toml
[project]
name = "energy-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "django>=6.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.0",
]
